feat: move eventType <-> extensions functions to new eventtypes.ts, add Field to eventType conversion functions
